+++ CHANGELOG.md
@@ -10,14 +10,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
    Unreleased section, uncommenting the header as necessary.
 -->
 
-## Unreleased
+<!-- ## Unreleased -->
 <!-- ### Added -->
+<!-- ### Removed -->
+<!-- ### Fixed -->
+<!-- ### Changed -->
+
+## [2.2.1] - 2019-07-23
 ### Changed
 * Elements should now override the new `_getUpdateComplete` method instead of the `updateComplete` getter, for compatibility with TypeScript ES5 output, which does not support calling a superclass getter (e.g.`super.updateComplete.then(...)`) due to [TypeScript#338](https://github.com/microsoft/TypeScript/issues/338).
-<!-- ### Removed -->
 ### Fixed
 * Fixed compatibility with Closure JS Compiler optimizations relating to static properties ([#732](https://github.com/Polymer/lit-element/issues/732)).
-
 
 ## [2.2.0] - 2019-06-11
 ### Added